Initial implementation
Impliments a leaky bucket rate limiter, that unlike
prorate continues to to count requests against the
limit even when the rate limiter is in the blocking
state.

This means that the client has to slow down, or
they will remain blocked indefinately.

Optionally a penalty can be added, that adds additonal tokens
to the bucket at the point that the limit is breached, to
futher ensure that the block lasts longer for clients that
are only marginly breaching the rate limit.